Import some code and add the ssh gerrit cmd logic
This is the first commit to start the new little project.
This change introduces a little (simple) way to dynamically
config the gerrit bot and a first python file which implements
the logic of running simple commands against a gerrit instance.
It's also able to provide a simple summary of the last change
of a given PS.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>