Import the first cephbot structure and its config
This change imports the first piece of code containing the cephbot
structure. It depends on both daemon and irc python modules and
implements the basic functions provided by the super class irc
SingleServerIRCBot.
This patch also adds some logging and provides a parameter to
properly configure it via log_path.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>