Handle pub msg(s) the same way of privmsg(s)
This change adds the logic of handling both priv and pub msgs
using the related, implemented callback if available.
If the msg cannot be processed (or the callback cannot be
discovered/executed), usage() is executed, informing people
about the next steps to interact w/ this bot.
In addition, a on_help callback is registered.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>