Make the bot able to retrieve gerrit results
This change brings a big improvement: the bot is now able
to run several gerrit subcommands (using the right syntax)
and query a specific submission_id.
In addition, a send command is implemented for the bot,
because it needs to wrap a msg on several lines (w/o run
into \n issues when privmsg is invoked).

Signed-off-by: Francesco Pantano <fpantano@redhat.com>