Support raw data print instead of prettytable
When the bot returns the values provided by a gerrit query (e.g.
ci logs), we need to return raw data instead of pretty table,
because it's faster and optimized for most of the clients.
This change just adds more logic within the summary/status
functions and returns the raw data of the query run against a
given submission_id.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>