Add 'pending_ceph' subcommand for gerrit queries
Now gerrit is able to resolve the 'pending_ceph' command using a new
config option. This is really helpful for people who needs to easily
interact with this bot w/o remembering the review id.
The helper is also improved and it's now able to suggest the proper
syntax when operators need to access to this new subcommand.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>